feat: merge chat page with webapp
