--- apps/cloud/src/app/features/chat/chat-input/chat-input.component.scss
@@ -1,7 +1,0 @@
-:host {
-  @apply w-full flex flex-col items-stretch lg:w-[800px] max-w-full;
-}
-
-.btn-canvas.active {
-  @apply bg-lime-50 text-lime-800;
-}
