+++ apps/cloud/src/app/features/chat/chat.service.ts
@@ -9,14 +9,16 @@ import { IXpert } from '../../@core'
 import { ToastrService } from '../../@core/services'
 import { ChatService } from '../../xpert/'
 import { ChatHomeService } from './home.service'
+import { Router } from '@angular/router'
 
 @Injectable()
 export class ChatPlatformService extends ChatService {
   readonly #translate = inject(TranslateService)
   readonly homeService = inject(ChatHomeService)
   readonly #toastr = inject(ToastrService)
+  readonly #router = inject(Router)
   readonly #location = inject(Location)
-  readonly paramRole = injectParams('role')
+  readonly paramRole = injectParams('name')
   readonly paramId = injectParams('id')
 
   readonly xperts = this.homeService.xperts
@@ -85,5 +87,13 @@ export class ChatPlatformService extends ChatService {
   async newConversation(xpert?: IXpert) {
     await super.newConversation()
     this.xpert$.next(xpert)
+  }
+
+  newConv(slug?: string) {
+    if (slug) {
+      this.#router.navigate(['/chat/x', slug])
+    } else {
+      this.#router.navigate(['/chat'])
+    }
   }
 }
