--- apps/cloud/src/app/features/chat/sidenav-menu/sidenav-menu.component.html
@@ -1,21 +1,0 @@
-<span class="flex" >
-  <button [matTooltip]="sidenav().opened ? (
-    'PAC.Chat.CloseSidebar' | translate: {Default: 'Close sidebar'}
-  ) : (
-    'PAC.Chat.OpenSidebar' | translate: {Default: 'Open sidebar'}
-  )"
-    class="pac-chat-button"
-    (click)="sidenav().toggle()"
-  >
-    <pac-chat-sidemenu-svg />
-  </button>
-</span>
-
-<span class="flex" >
-  <button [matTooltip]="'PAC.Chat.NewChat' | translate: {Default: 'New chat'}"
-    class="pac-chat-button"
-    (click)="newConversation()"
-  >
-    <pac-chat-new-chat-svg />
-  </button>
-</span>
