--- apps/cloud/src/app/features/chat/sidenav-menu/sidenav-menu.component.scss
@@ -1,3 +1,0 @@
-:host {
-  @apply flex justify-between items-center;
-}
