--- apps/cloud/src/app/features/chat/toolbar/toolbar.component.scss
@@ -1,3 +1,0 @@
-:host {
-    @apply flex justify-start items-center p-2;
-}
