+++ apps/cloud/src/app/xpert/canvas/canvas.component.scss
