+++ apps/cloud/src/app/xpert/canvas/computer/computer.component.scss
