+++ apps/cloud/src/app/xpert/canvas/dashboard/dashboard.component.scss
