+++ apps/cloud/src/app/xpert/canvas/file-viewer/file-viewer.component.scss
