+++ apps/cloud/src/app/xpert/chat-input/chat-input.component.scss
@@ -1,8 +1,7 @@
 :host {
-  @apply w-full flex flex-col items-stretch lg:w-[800px] max-w-full pt-4 lg:pt-8;
+  @apply w-full flex flex-col items-stretch lg:w-[800px] max-w-full;
 }
 
-// Copilot
-.ngm-colpilot__input.mat-mdc-input-element {
-  @apply outline-none bg-transparent;
+.btn-canvas.active {
+  @apply bg-lime-50 text-lime-800;
 }
