+++ apps/cloud/src/app/xpert/conversation/conversation.component.html
@@ -27,7 +27,7 @@
   }
 } @empty {
   @if (!loadingConv() && xpert()) {
-    <div class="w-full h-full flex flex-col justify-center items-center gap-4">
+    <div class="flex-1 w-full h-full flex flex-col justify-center items-center gap-4">
       <emoji-avatar [avatar]="xpert().avatar" large class="rounded-xl overflow-hidden shadow-sm" />
 
       <div class="w-full flex flex-col items-center gap-2">
@@ -41,6 +41,7 @@
             <button class="relative flex w-40 flex-col gap-2 rounded-xl border border-token-border-light px-2 pb-3 pt-2 text-start align-top text-sm shadow-sm
               md:text-base md:px-3 md:pt-3 md:pb-4 md:rounded-2xl
               transition enabled:hover:bg-token-main-surface-secondary disabled:cursor-not-allowed"
+              (click)="onChat(statement)"
             >
               <div class="line-clamp-3 max-w-full text-balance font-light text-neutral-500 dark:text-neutral-500 break-word"
                 [title]="statement"
