+++ apps/cloud/src/app/xpert/home.service.ts
@@ -24,6 +24,11 @@ export class XpertHomeService {
   readonly lang = this.appService.lang
 
   readonly conversations = signal<IChatConversation[]>([])
+  readonly currentPage = signal(0)
+  readonly pagesCompleted = signal(false)
+  /**
+   * The ID of the current conversation
+   */
   readonly conversationId = signal<string>(null)
 
   readonly conversation = signal<IChatConversation>(null)
