+++ apps/cloud/src/app/xpert/home/home.component.scss
@@ -1,5 +1,5 @@
 :host {
-  background-color: var(--ngm-color-surface);
+  background-color: var(--ngm-shell-primary-container-bg-color);
   @apply max-w-full flex-1 flex;
 }
