+++ apps/cloud/src/app/xpert/index.ts
@@ -4,3 +4,5 @@ export * from './chat-input/chat-input.component'
 export * from './home.service'
 export * from './ocap.service'
 export * from './component-message/index'
+export * from './xpert/xpert.component'
+export * from './conversations/conversations.component'
